Refactoring of Datum / SmartType
Datum and SmartType are now seperate classes. An instance of Datum (or a
subclass) only represents a table row, while the SmartType defines the
Schema. It cuts away most of the overhead which is currently present for
creatings Datum instances. Unfortunately, it breaks the current API
which requires changing current SmartType definitions. While it makes
setting up object definitions a bit more complicated, it simplifies the
construct logically.